# backend/monolith/swagger-spec.json
{"openapi":"3.0.0","info":{"title":"API","description":"","version":"7.0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateCmpBookingDto":{"type":"object","properties":{"companyId":{"type":"string"},"staffId":{"type":"string"},"userId":{"type":"string"},"serviceId":{"type":"string"}}},"UserDto":{"type":"object","properties":{"name":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"bookings":{"type":"array","items":{"type":"string"}},"favoritesUserServices":{"type":"array","items":{"type":"string"}}}},"BookingDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"}}},"UpdateCmpBookingDto":{"type":"object","properties":{"companyId":{"type":"string"},"staffId":{"type":"string"},"userId":{"type":"string"},"serviceId":{"type":"string"}}},"CompanyLoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CompanyDto":{"type":"object","properties":{"fullName":{"type":"string"},"email":{"type":"string"}}},"TokenPayloadDto":{"type":"object","properties":{"expiresIn":{"type":"number"},"accessToken":{"type":"string"}},"required":["expiresIn","accessToken"]},"LoginPayloadDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"},"token":{"$ref":"#/components/schemas/TokenPayloadDto"}},"required":["user","token"]},"CreateCmpCompanyDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","minLength":6},"fullName":{"type":"string","minLength":6}},"required":["email","password","fullName"]},"UpdateCmpCompanyDto":{"type":"object","properties":{"fullName":{"type":"string"},"email":{"type":"string"}}},"CreateCmpServiceDto":{"type":"object","properties":{"name":{"type":"string"}}},"ServiceDto":{"type":"object","properties":{"name":{"type":"string"},"branches":{"type":"array","items":{"type":"string"}}}},"UpdateCmpServiceDto":{"type":"object","properties":{"name":{"type":"string"}}},"CreateCmpStaffDto":{"type":"object","properties":{"name":{"type":"string"}}},"StaffDto":{"type":"object","properties":{"name":{"type":"string"}}},"UpdateCmpStaffDto":{"type":"object","properties":{"name":{"type":"string"}}},"BranchDto":{"type":"object","properties":{"name":{"type":"string"},"services":{"type":"array","items":{"type":"string"}}}},"PageMetaDto":{"type":"object","properties":{"page":{"type":"number"},"take":{"type":"number"},"itemCount":{"type":"number"},"pageCount":{"type":"number"},"hasPreviousPage":{"type":"boolean"},"hasNextPage":{"type":"boolean"}},"required":["page","take","itemCount","pageCount","hasPreviousPage","hasNextPage"]},"PageDto":{"type":"object","properties":{"data":{"type":"array","items":{"type":"array"}},"meta":{"$ref":"#/components/schemas/PageMetaDto"}},"required":["data","meta"]},"UpdateAppUserDto":{"type":"object","properties":{"name":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"CreateAppReviewDto":{"type":"object","properties":{"name":{"type":"string"}}},"ReviewDto":{"type":"object","properties":{"name":{"type":"string"}}},"UpdateAppReviewDto":{"type":"object","properties":{"name":{"type":"string"}}},"UserLoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UserRegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","minLength":6},"name":{"type":"string","minLength":3},"lastName":{"type":"string","minLength":3}},"required":["email","password","name","lastName"]}}},"paths":{"/company/bookings":{"post":{"operationId":"CmpBookingController_createBooking","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCmpBookingDto"}}}},"responses":{"200":{"description":"Successfully created booking","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/bookings"],"security":[{"bearer":[]}]},"put":{"operationId":"CmpBookingController_updateBooking","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCmpBookingDto"}}}},"responses":{"200":{"description":"Successfully updated booking","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/bookings"],"security":[{"bearer":[]}]},"get":{"operationId":"CmpBookingController_getCompanyBookings","parameters":[],"responses":{"200":{"description":"Get bookings list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/bookings"],"security":[{"bearer":[]}]}},"/company/bookings/{id}":{"get":{"operationId":"CmpBookingController_getCompanyBooking","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get booking","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/bookings"],"security":[{"bearer":[]}]},"delete":{"operationId":"CmpBookingController_deleteCompanyBooking","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete booking","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/bookings"],"security":[{"bearer":[]}]}},"/company/auth/login":{"post":{"operationId":"CmpAuthController_companyLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyLoginDto"}}}},"responses":{"200":{"description":"Company info with access token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginPayloadDto"}}}}},"tags":["company/auth"]}},"/company/auth/register":{"post":{"operationId":"CmpAuthController_companyRegister","parameters":[{"schema":{}}],"requestBody":{"required":false,"content":{"multipart/form-data":{"schema":{"allOf":[{"$ref":"#/components/schemas/CreateCmpCompanyDto"},{"properties":{"avatar":{"type":"string","format":"binary"}},"type":"object"}]}}}},"responses":{"200":{"description":"Successfully Registered","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}}},"tags":["company/auth"]}},"/company/auth/me":{"get":{"operationId":"CmpAuthController_getCurrentCompany","parameters":[],"responses":{"200":{"description":"current app-company-module-app-company info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}}},"tags":["company/auth"],"security":[{"bearer":[]}]}},"/company/companies":{"post":{"operationId":"CmpCompanyController_createCompany","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCmpCompanyDto"}}}},"responses":{"200":{"description":"Successfully created company","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/companies"],"security":[{"bearer":[]}]},"put":{"operationId":"CmpCompanyController_updateCompany","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCmpCompanyDto"}}}},"responses":{"200":{"description":"Successfully updated company","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/companies"],"security":[{"bearer":[]}]},"get":{"operationId":"CmpCompanyController_getCompanies","parameters":[],"responses":{"200":{"description":"Get companies list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/companies"],"security":[{"bearer":[]}]}},"/company/companies/{id}":{"get":{"operationId":"CmpCompanyController_getCompany","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get company","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/companies"],"security":[{"bearer":[]}]},"delete":{"operationId":"CmpCompanyController_deleteCompany","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete company","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/companies"],"security":[{"bearer":[]}]}},"/company/services":{"post":{"operationId":"CmpServiceController_createService","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCmpServiceDto"}}}},"responses":{"200":{"description":"Successfully created service","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/services"],"security":[{"bearer":[]}]},"put":{"operationId":"CmpServiceController_updateService","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCmpServiceDto"}}}},"responses":{"200":{"description":"Successfully updated service","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/services"],"security":[{"bearer":[]}]},"get":{"operationId":"CmpServiceController_getCompanyServices","parameters":[],"responses":{"200":{"description":"Get services list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/services"],"security":[{"bearer":[]}]}},"/company/services/{id}":{"get":{"operationId":"CmpServiceController_getCompanyService","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get service","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/services"],"security":[{"bearer":[]}]},"delete":{"operationId":"CmpServiceController_deleteCompanyService","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete service","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/services"],"security":[{"bearer":[]}]}},"/company/staffs":{"post":{"operationId":"CmpStaffController_createStaff","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCmpStaffDto"}}}},"responses":{"200":{"description":"Successfully created staff","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StaffDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/staffs"],"security":[{"bearer":[]}]},"put":{"operationId":"CmpStaffController_updateStaff","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCmpStaffDto"}}}},"responses":{"200":{"description":"Successfully updated staff","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StaffDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/staffs"],"security":[{"bearer":[]}]},"get":{"operationId":"CmpStaffController_getCompanyStaffs","parameters":[],"responses":{"200":{"description":"Get staffs list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StaffDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/staffs"],"security":[{"bearer":[]}]}},"/company/staffs/{id}":{"get":{"operationId":"CmpStaffController_getCompanyStaff","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get staff","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StaffDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/staffs"],"security":[{"bearer":[]}]},"delete":{"operationId":"CmpStaffController_deleteCompanyStaff","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete staff","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StaffDto"}}}},"401":{"description":"Unauthorized"}},"tags":["company/staffs"],"security":[{"bearer":[]}]}},"/app/bookings":{"get":{"operationId":"AppBookingController_getBookings","parameters":[],"responses":{"200":{"description":"Get bookings list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/bookings"],"security":[{"bearer":[]}]},"post":{"operationId":"AppBookingController_createBooking","parameters":[],"responses":{"200":{"description":"Create booking","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/bookings"],"security":[{"bearer":[]}]}},"/app/branches":{"get":{"operationId":"AppBranchController_getBranches","parameters":[],"responses":{"200":{"description":"Get branches list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BranchDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/branches"],"security":[{"bearer":[]}]}},"/app/companies":{"get":{"operationId":"AppCompanyController_getCompanies","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":10,"type":"number"}},{"name":"q","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Get companies list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageDto"}}}}},"tags":["app/companies"]}},"/app/companies/{id}":{"get":{"operationId":"AppCompanyController_getCompany","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get companies list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/companies"],"security":[{"bearer":[]}]}},"/app/services":{"get":{"operationId":"AppServiceController_getServices","parameters":[],"responses":{"200":{"description":"Get services list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/services"],"security":[{"bearer":[]}]}},"/app/services/favorites":{"get":{"operationId":"AppServiceController_getFavoriteServices","parameters":[],"responses":{"200":{"description":"Get favorite services","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/services"],"security":[{"bearer":[]}]}},"/app/services/{id}":{"get":{"operationId":"AppServiceController_getService","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get service","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/services"],"security":[{"bearer":[]}]}},"/app/users/{id}":{"get":{"operationId":"AppUserController_getUsers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get users list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/users"],"security":[{"bearer":[]}]},"delete":{"operationId":"AppUserController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully deleted user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/users"],"security":[{"bearer":[]}]}},"/app/users":{"put":{"operationId":"AppUserController_updateUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAppUserDto"}}}},"responses":{"200":{"description":"Successfully updated user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/users"],"security":[{"bearer":[]}]}},"/app/reviews":{"post":{"operationId":"AppReviewController_createReview","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAppReviewDto"}}}},"responses":{"200":{"description":"Successfully created review","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/reviews"],"security":[{"bearer":[]}]},"put":{"operationId":"AppReviewController_updateReview","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAppReviewDto"}}}},"responses":{"200":{"description":"Successfully updated review","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/reviews"],"security":[{"bearer":[]}]},"get":{"operationId":"AppReviewController_getUserReviews","parameters":[],"responses":{"200":{"description":"Get reviews list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/reviews"],"security":[{"bearer":[]}]}},"/app/reviews/{id}":{"get":{"operationId":"AppReviewController_getUserReview","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get review","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/reviews"],"security":[{"bearer":[]}]},"delete":{"operationId":"AppReviewController_deleteUserReview","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete review","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/reviews"],"security":[{"bearer":[]}]}},"/app/auth/login":{"post":{"operationId":"AppUserAuthController_userLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}}},"responses":{"200":{"description":"User info with access token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginPayloadDto"}}}}},"tags":["app/auth"]}},"/app/auth/register":{"post":{"operationId":"AppUserAuthController_userRegister","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisterDto"}}}},"responses":{"200":{"description":"Successfully Registered","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["app/auth"]}},"/app/auth/me":{"get":{"operationId":"AppUserAuthController_getCurrentUser","parameters":[],"responses":{"200":{"description":"current user info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"Unauthorized"}},"tags":["app/auth"],"security":[{"bearer":[]}]}}}}